Rename window_base.* to window.*
