+++ simulant/event_listener.cpp
@@ -1,6 +1,6 @@
 
 #include "event_listener.h"
-#include "window_base.h"
+#include "window.h"
 
 namespace smlt {
