+++ simulant/idle_task_manager.cpp
@@ -21,7 +21,7 @@
 #include <functional>
 #include "deps/kazlog/kazlog.h"
 #include "idle_task_manager.h"
-#include "window_base.h"
+#include "window.h"
 #include "utils/gl_thread_check.h"
 
 namespace smlt {
