+++ simulant/kos_window.h
@@ -4,7 +4,7 @@
 
 #include <kos.h>
 
-#include "window_base.h"
+#include "window.h"
 
 namespace smlt {
