+++ simulant/managers.cpp
@@ -20,7 +20,7 @@
 
 #include "managers.h"
 #include "background.h"
-#include "window_base.h"
+#include "window.h"
 #include "stage.h"
 #include "nodes/camera.h"
 #include "render_sequence.h"
