+++ simulant/panels/partitioner_panel.cpp
@@ -19,7 +19,7 @@
 
 
 #include "partitioner_panel.h"
-#include "../window_base.h"
+#include "../window.h"
 #include "../stage.h"
 #include "../nodes/actor.h"
 #include "../partitioner.h"
