+++ simulant/virtual_gamepad.cpp
@@ -20,7 +20,7 @@
 #include "deps/kazlog/kazlog.h"
 #include "utils/unicode.h"
 #include "virtual_gamepad.h"
-#include "window_base.h"
+#include "window.h"
 #include "stage.h"
 #include "render_sequence.h"
 #include "nodes/ui/ui_manager.h"
